# docs/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "structured-search API",
    "description": "Task-scoped HTTP API for structured, auditable search workflows.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/tasks/{task_id}/profiles": {
      "get": {
        "summary": "List available profiles for one task",
        "operationId": "get_profiles_v1_tasks__task_id__profiles_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileSummary"
                  },
                  "title": "Response Get Profiles V1 Tasks  Task Id  Profiles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/profiles/{profile_id}/bundle": {
      "get": {
        "summary": "Load a task/profile bundle",
        "operationId": "get_bundle_v1_tasks__task_id__profiles__profile_id__bundle_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Validate and save a task/profile bundle",
        "operationId": "put_bundle_v1_tasks__task_id__profiles__profile_id__bundle_put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileBundle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/prompt/generate": {
      "post": {
        "summary": "Compose an extraction prompt for a task/profile and step",
        "operationId": "post_generate_prompt_v1_tasks__task_id__prompt_generate_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/jsonl/validate": {
      "post": {
        "summary": "Parse JSONL tolerantly and validate records against task schema",
        "operationId": "post_jsonl_validate_v1_tasks__task_id__jsonl_validate_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonlValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonlValidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/run/validate": {
      "post": {
        "summary": "Dry-run validation for run prerequisites without scoring",
        "operationId": "post_task_run_validate_v1_tasks__task_id__run_validate_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunValidateSummary"
                }
              }
            }
          },
          "404": {
            "description": "Task/profile not found"
          },
          "422": {
            "description": "Validation/config error"
          }
        }
      }
    },
    "/v1/tasks/{task_id}/run": {
      "post": {
        "summary": "Score pre-validated records using task profile runtime config",
        "operationId": "post_task_run_v1_tasks__task_id__run_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunScoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task/profile not found"
          },
          "422": {
            "description": "Validation/config error"
          },
          "500": {
            "description": "Run failed"
          }
        }
      }
    },
    "/v1/tasks/{task_id}/actions/gen-cv": {
      "post": {
        "summary": "Generate a CV markdown and structured JSON",
        "operationId": "post_action_gen_cv_v1_tasks__task_id__actions_gen_cv_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenCVRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenCVResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task/profile not found"
          },
          "422": {
            "description": "Validation error in request payload"
          },
          "503": {
            "description": "CV provider unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BundleWriteResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "BundleWriteResponse",
        "description": "HTTP response body for PUT bundle endpoint."
      },
...
      "GenCVRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "job": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job"
          },
          "candidate_profile": {
            "$ref": "#/components/schemas/CvCandidateInput"
          },
          "selected_claim_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Claim Ids"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "allow_mock_fallback": {
            "type": "boolean",
            "title": "Allow Mock Fallback",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "job",
          "candidate_profile"
        ],
        "title": "GenCVRequest",
        "description": "Request body for POST /v1/tasks/{task_id}/actions/gen-cv."
      },
      "GenCVResponse": {
        "properties": {
          "cv_markdown": {
            "type": "string",
            "title": "Cv Markdown"
          },
          "generated_cv_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Cv Json"
          },
          "model_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Info"
          }
        },
        "type": "object",
        "required": [
          "cv_markdown"
        ],
        "title": "GenCVResponse",
        "description": "Response body for gen-cv action."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonlValidateRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "raw_jsonl": {
            "type": "string",
            "title": "Raw Jsonl"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "raw_jsonl"
        ],
        "title": "JsonlValidateRequest",
        "description": "Request body for POST JSONL validation."
      },
      "JsonlValidateResponse": {
        "properties": {
          "valid_records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Valid Records"
          },
          "invalid_records": {
            "items": {
              "$ref": "#/components/schemas/JsonlInvalidRecord"
            },
            "type": "array",
            "title": "Invalid Records"
          },
          "metrics": {
            "$ref": "#/components/schemas/JsonlValidateMetrics"
          }
        },
        "type": "object",
        "required": [
          "valid_records",
          "invalid_records",
          "metrics"
        ],
        "title": "JsonlValidateResponse",
        "description": "HTTP response body for JSONL validation."
      },
      "ProfileBundle": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "default": "job_search"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "task": {
            "additionalProperties": true,
            "type": "object",
            "title": "Task"
          },
          "task_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Task Config"
          },
          "user_profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Profile"
          },
          "domain_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Schema"
          },
          "result_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Schema"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "constraints",
          "task",
          "task_config"
        ],
        "title": "ProfileBundle",
        "description": "All editable config for one task/profile."
      },
      "ProfileSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "updated_at"
        ],
        "title": "ProfileSummary",
        "description": "Summary item from GET /v1/tasks/{task_id}/profiles."
      },
      "PromptGenerateRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "step": {
            "type": "string",
            "title": "Step",
            "default": "S3_execute"
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "PromptGenerateRequest",
        "description": "Request body for POST prompt generation."
      },
      "PromptGenerateResponse": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "step": {
            "type": "string",
            "title": "Step"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "constraints_embedded": {
            "type": "boolean",
            "title": "Constraints Embedded"
          },
          "prompt_hash": {
            "type": "string",
            "title": "Prompt Hash"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "step",
          "prompt",
          "constraints_embedded",
          "prompt_hash"
        ],
        "title": "PromptGenerateResponse",
        "description": "HTTP response body for POST prompt generation."
      },
...
      "RunScoreRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Records"
          },
          "require_snapshot": {
            "type": "boolean",
            "title": "Require Snapshot",
            "description": "If true, snapshot persistence is mandatory. Run fails when snapshot write fails.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "records"
        ],
        "title": "RunScoreRequest",
        "description": "Request body for POST run."
      },
      "RunScoreResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "scored_records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Scored Records"
          },
          "metrics": {
            "$ref": "#/components/schemas/RunResponseMetrics"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors"
          },
          "snapshot_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Dir"
          },
          "snapshot_status": {
            "type": "string",
            "enum": [
              "written",
              "failed"
            ],
            "title": "Snapshot Status",
            "default": "written"
          },
          "snapshot_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "profile_id",
          "scored_records",
          "metrics",
          "errors"
        ],
        "title": "RunScoreResponse",
        "description": "HTTP response body for POST run."
      },
...
      "RunValidateSummary": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "total_records": {
            "type": "integer",
            "title": "Total Records"
          },
          "valid_records": {
            "type": "integer",
            "title": "Valid Records"
          },
          "invalid_records": {
            "type": "integer",
            "title": "Invalid Records"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/IngestError"
            },
            "type": "array",
            "title": "Errors"
          },
          "checks": {
            "$ref": "#/components/schemas/RunValidateChecks"
          },
          "snapshot_probe_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Probe Dir"
          },
          "snapshot_probe_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Probe Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "profile_id",
          "total_records",
          "valid_records",
          "invalid_records",
          "errors",
          "checks"
        ],
        "title": "RunValidateSummary",
        "description": "Dry-run summary for validating whether /run can execute."
      },
...
    }
  }
}